+++ src/app/js/annotation/AnnotationDisabledIcon.js
@@ -1,61 +1,45 @@
-import { Box, Tooltip, useTheme } from '@mui/material';
+import PropTypes from 'prop-types';
 import React from 'react';
-import { useTranslate } from '../i18n/I18NContext';
 
-export function AnnotationDisabledIcon() {
-    const theme = useTheme();
-    const { translate } = useTranslate();
+export function AnnotationDisabledIcon({ fill = '#000', ...props } = {}) {
     return (
-        <Tooltip title={translate('annotation_disabled_tooltip')}>
-            <Box
-                sx={{
-                    width: 20,
-                    height: 20,
-                    display: 'flex',
-                    alignItems: 'center',
-                    justifyContent: 'center',
-                    marginTop: -0.5,
-                }}
+        <svg
+            xmlns="http://www.w3.org/2000/svg"
+            width="20"
+            height="18"
+            viewBox="0 0 23 21"
+            fill={fill}
+            {...props}
+        >
+            <path
+                d="M17.3478 4.34184C15.7788 2.84798 13.6982 2.01024 11.5318 2.00009C9.36545 1.98994 7.27709 2.80815 5.69418 4.28725C4.11127 5.76635 3.15351 7.79449 3.01691 9.95659C2.88031 12.1187 3.57521 14.2512 4.95938 15.9178C5.05789 16.0427 5.13 16.1863 5.17135 16.3399C5.21269 16.4934 5.22241 16.6538 5.1999 16.8113L4.97267 18.6347C4.94803 18.8342 4.96591 19.0367 5.02515 19.2288C5.08439 19.4209 5.18364 19.5983 5.31636 19.7493C5.44908 19.9003 5.61227 20.0214 5.79519 20.1048C5.97811 20.1882 6.17662 20.2319 6.37765 20.2331C6.57128 20.2325 6.76279 20.1927 6.94058 20.116L9.57647 18.9863C9.80129 18.8954 10.047 18.8688 10.286 18.9095C11.4594 19.0808 12.6557 19.0022 13.7966 18.6789C15.4041 18.2205 16.8421 17.3011 17.9328 16.0343C19.0235 14.7675 19.7191 13.2089 19.9335 11.5511C20.1 10.2307 19.9527 8.88972 19.5034 7.63704C19.0541 6.38436 18.3155 5.25538 17.3478 4.34184Z"
+                fill={fill}
+            />
+            <mask
+                id="path-2-outside-1_698_465"
+                maskUnits="userSpaceOnUse"
+                x="-0.152344"
+                y="0"
+                width="23"
+                height="21"
+                fill="black"
             >
-                <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    width="20"
-                    height="18"
-                    viewBox="0 0 23 21"
-                    fill="none"
-                >
-                    <path
-                        d="M17.3478 4.34184C15.7788 2.84798 13.6982 2.01024 11.5318 2.00009C9.36545 1.98994 7.27709 2.80815 5.69418 4.28725C4.11127 5.76635 3.15351 7.79449 3.01691 9.95659C2.88031 12.1187 3.57521 14.2512 4.95938 15.9178C5.05789 16.0427 5.13 16.1863 5.17135 16.3399C5.21269 16.4934 5.22241 16.6538 5.1999 16.8113L4.97267 18.6347C4.94803 18.8342 4.96591 19.0367 5.02515 19.2288C5.08439 19.4209 5.18364 19.5983 5.31636 19.7493C5.44908 19.9003 5.61227 20.0214 5.79519 20.1048C5.97811 20.1882 6.17662 20.2319 6.37765 20.2331C6.57128 20.2325 6.76279 20.1927 6.94058 20.116L9.57647 18.9863C9.80129 18.8954 10.047 18.8688 10.286 18.9095C11.4594 19.0808 12.6557 19.0022 13.7966 18.6789C15.4041 18.2205 16.8421 17.3011 17.9328 16.0343C19.0235 14.7675 19.7191 13.2089 19.9335 11.5511C20.1 10.2307 19.9527 8.88972 19.5034 7.63704C19.0541 6.38436 18.3155 5.25538 17.3478 4.34184Z"
-                        fill={theme.palette.grey[500]}
-                    />
-                    <mask
-                        id="path-2-outside-1_698_465"
-                        maskUnits="userSpaceOnUse"
-                        x="-0.152344"
-                        y="0"
-                        width="23"
-                        height="21"
-                        fill="black"
-                    >
-                        <rect
-                            fill="white"
-                            x="-0.152344"
-                            width="23"
-                            height="21"
-                        />
-                        <path d="M3.00004 2L20.829 18.4828L19.6766 19.5809L1.84766 3.09806L3.00004 2Z" />
-                    </mask>
-                    <path
-                        d="M3.00004 2L20.829 18.4828L19.6766 19.5809L1.84766 3.09806L3.00004 2Z"
-                        fill={theme.palette.grey[500]}
-                    />
-                    <path
-                        d="M3.00004 2L3.67889 1.26572L2.9897 0.628569L2.31021 1.27604L3.00004 2ZM20.829 18.4828L21.5188 19.2068L22.2902 18.4718L21.5078 17.7485L20.829 18.4828ZM19.6766 19.5809L18.9978 20.3152L19.687 20.9523L20.3665 20.3048L19.6766 19.5809ZM1.84766 3.09806L1.15782 2.3741L0.386477 3.10908L1.16881 3.83235L1.84766 3.09806ZM2.3212 2.73428L20.1502 19.2171L21.5078 17.7485L3.67889 1.26572L2.3212 2.73428ZM20.1392 17.7588L18.9868 18.8569L20.3665 20.3048L21.5188 19.2068L20.1392 17.7588ZM20.3555 18.8466L2.5265 2.36378L1.16881 3.83235L18.9978 20.3152L20.3555 18.8466ZM2.53749 3.82203L3.68988 2.72396L2.31021 1.27604L1.15782 2.3741L2.53749 3.82203Z"
-                        fill="white"
-                        mask="url(#path-2-outside-1_698_465)"
-                    />
-                </svg>
-            </Box>
-        </Tooltip>
+                <rect fill="white" x="-0.152344" width="23" height="21" />
+                <path d="M3.00004 2L20.829 18.4828L19.6766 19.5809L1.84766 3.09806L3.00004 2Z" />
+            </mask>
+            <path
+                d="M3.00004 2L20.829 18.4828L19.6766 19.5809L1.84766 3.09806L3.00004 2Z"
+                fill={fill}
+            />
+            <path
+                d="M3.00004 2L3.67889 1.26572L2.9897 0.628569L2.31021 1.27604L3.00004 2ZM20.829 18.4828L21.5188 19.2068L22.2902 18.4718L21.5078 17.7485L20.829 18.4828ZM19.6766 19.5809L18.9978 20.3152L19.687 20.9523L20.3665 20.3048L19.6766 19.5809ZM1.84766 3.09806L1.15782 2.3741L0.386477 3.10908L1.16881 3.83235L1.84766 3.09806ZM2.3212 2.73428L20.1502 19.2171L21.5078 17.7485L3.67889 1.26572L2.3212 2.73428ZM20.1392 17.7588L18.9868 18.8569L20.3665 20.3048L21.5188 19.2068L20.1392 17.7588ZM20.3555 18.8466L2.5265 2.36378L1.16881 3.83235L18.9978 20.3152L20.3555 18.8466ZM2.53749 3.82203L3.68988 2.72396L2.31021 1.27604L1.15782 2.3741L2.53749 3.82203Z"
+                fill="white"
+                mask="url(#path-2-outside-1_698_465)"
+            />
+        </svg>
     );
 }
+
+AnnotationDisabledIcon.propTypes = {
+    fill: PropTypes.string,
+};
